% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ORs_CIs}
\alias{get_ORs_CIs}
\title{Return conf. intervals for Odds Ratios of GLM}
\usage{
get_ORs_CIs(x)
}
\arguments{
\item{x}{A model of class}
}
\value{
A table of odds ratios and 95% CI
}
\description{
Return conf. intervals for Odds Ratios of GLM
}
\examples{
\dontrun{
get_ORs_CIs(mod1)
}
}
