% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ORs_CIs_alt}
\alias{get_ORs_CIs_alt}
\title{Return conf. intervals for Odds Ratios of GLM (alt version)}
\usage{
get_ORs_CIs_alt(x)
}
\arguments{
\item{x}{A model of class}
}
\value{
A table of odds ratios and 95% CI
}
\description{
Unclear why this alternative version exists...
}
\examples{
\dontrun{
get_ORs_CIs_alt(mod1)
}
}
