% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_binom}
\alias{inf_binom}
\title{Run binomial GLM for informal consultation quality components}
\usage{
inf_binom(x, y)
}
\arguments{
\item{x}{The "informal" data.frame}

\item{y}{The response variable}
}
\value{
A list of results including:
  \describe{
    \item{mods}{Four GLM (binomial) model objects}
    \item{AICs}{AICc values for the four models}
    \item{summaries}{Results from \code{summary(mod)} for the four models}
    \item{aovs}{Analysis of Variance for the four models}
  }
}
\description{
Run binomial GLM for informal consultation quality components
}
\examples{
inf_binom(informal, informal$map)
}

