% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_glm_combo_informal}
\alias{run_glm_combo_informal}
\title{Run binomial GLM for overall quality of informal consultations}
\usage{
run_glm_combo_informal(x)
}
\arguments{
\item{x}{The 'informal' data.frame of just informal consultations}
}
\value{
A list of results including:
  \describe{
    \item{mods}{Nine GLM (binomial) model objects}
    \item{AICs}{AICc values for the nine models}
    \item{summaries}{Results from \code{summary(mod)} for the nine models}
    \item{aovs}{Analysis of Variance for the nine models}
  }
}
\description{
Run binomial GLM for overall quality of informal consultations
}
\examples{
\dontrun{
run_glm_combo(combo)
}
}
