% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_glm_Ospec}
\alias{run_glm_Ospec}
\title{Run binomial GLM for overall consultation specificity}
\usage{
run_glm_Ospec(x)
}
\arguments{
\item{x}{The data.frame on which to run the nine candidate models}
}
\value{
A list of results including:
  \describe{
    \item{mods}{Nine GLM (binomial) model objects}
    \item{AICs}{AICc values for the nine models}
    \item{summaries}{Results from \code{summary(mod)} for the nine models}
    \item{aovs}{Analysis of Variance for the nine models}
  }
}
\description{
Run binomial GLM for overall consultation specificity
}
\examples{
run_glm_Ospec(formal)
}
