% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_polrs}
\alias{run_polrs}
\title{Run ordinal linear regression on consultation quality components}
\usage{
run_polrs(x, y)
}
\arguments{
\item{x}{The 'combo' data.frame of all consultations}

\item{y}{The response variable (quality component) for the models}
}
\value{
A list of results including:
  \describe{
    \item{mods}{Four GLM (binomial) model objects}
    \item{AICs}{AICc values for the four models}
    \item{summaries}{Results from \code{summary(mod)} for the four models}
    \item{aovs}{Analysis of Variance for the four models}
  }
}
\description{
Run ordinal linear regression on consultation quality components
}
\examples{
\dontrun{
run_polrs(formal, formal$status_Q)
}
}
